Split integration tests into separate classes
